app/requirements.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
djangorestframework==3.11.2
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
